// typings/discord/DiscordInviteType.ts
// Generated by https://quicktype.io

export interface DiscordInviteType {
  code: string;
  type: number;
  expires_at: null;
  guild: Guild;
  channel: Channel;
  inviter: Inviter;
}

export interface Channel {
  id: string;
  name: string;
  type: number;
}

export interface Guild {
  id: string;
  name: string;
  splash: null;
  banner: null;
  description: null;
  icon: string;
  features: string[];
  verification_level: number;
  vanity_url_code: null;
  premium_subscription_count: number;
  nsfw: boolean;
  nsfw_level: number;
}

export interface Inviter {
  id: string;
  username: string;
  avatar: string;
  avatar_decoration: null;
  discriminator: string;
  public_flags: number;
}

export interface UnknownInvite {
  message: "Unknown Invite";
  code: 10006;
}